reorg types
