+++ src/types.ts
@@ -309,77 +309,6 @@ export interface LocalTarget {
 	Id: string;
 }
 
-export interface XCMV2DestBenificiary {
-	V2: {
-		parents: string | number;
-		interior: {
-			X1: { AccountId32: { id: string } };
-		};
-	};
-}
-
-export interface XCMV3DestBenificiary {
-	V3: {
-		parents: string | number;
-		interior: {
-			X1: { AccountId32: { id: string } };
-		};
-	};
-}
-
-export interface XCMV2ParachainDestBenificiary {
-	V2: {
-		parents: string | number;
-		interior: {
-			X2: [{ Parachain: string }, { AccountId32: { id: string } }];
-		};
-	};
-}
-
-export interface XCMV3ParachainDestBenificiary {
-	V3: {
-		parents: string | number;
-		interior: {
-			X2: [{ Parachain: string }, { AccountId32: { id: string } }];
-		};
-	};
-}
-
-export type XCMDestBenificiary =
-	| XCMV3DestBenificiary
-	| XCMV2DestBenificiary
-	| XCMV2ParachainDestBenificiary
-	| XCMV3ParachainDestBenificiary;
-
-export interface XcmWeightUnlimited {
-	Unlimited: null | undefined;
-}
-
-export interface XcmWeightLimited {
-	Limited: {
-		refTime: string;
-		proofSize: string;
-	};
-}
-
-export type XcmWeight = XcmWeightUnlimited | XcmWeightLimited;
-
-export interface Args {
-	dest?: LocalDest;
-	beneficiary?: XCMDestBenificiary;
-	target?: LocalTarget;
-}
-
-export interface Method {
-	args: Args;
-	method: string;
-	section: string;
-}
-export interface SubmittableMethodData {
-	isSigned: boolean;
-	method: Method;
-}
-
 export type AssetInfo = {
 	id: string;
 	symbol: string;
